pymavlink: Cope with the system time moving backwards
A user noticed that if they move time back on their PC (due to ntp in this
case).  Then that can cause mavproxy/droneapi to drop the link.
The problem was that mavproxy stopped sending heartbeats because the timer
is now quite a bit back in time.  This change fixes the periodic_event
class to restart timers (and warn user) if this unusual case occurs.